consumer/requirements.txt
confluent-kafka>=2.3.0
confluent-kafka[avro]>=2.3.0
requests>=2.31.0
pydantic>=2.0.0
PyYAML>=6.0.1